+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+
+
+## [1.1.2] - 2026-01-06
+
+### Fixed
+- The plugin updates your brush name if you renamed it in Brush Editor
+
+### Changed
+- Preset Groups docker can be resized to an even smaller width
+- Brush Size textbox is smaller as it took extra unnecessary space
+
+
+
 ## [1.1.0] - 2025-12-30
 
 ### Changed
@@ -13,6 +26,8 @@ All notable changes to this project will be documented in this file.
 - Stuttering/lag while drawing, zooming, panning with the plugin active
 - Unnecessary CPU usage when the Preset Groups docker is hidden
 
+
+
 ## [1.0.8] - 2025-12-26
 
 ### Changed
@@ -24,12 +39,16 @@ All notable changes to this project will be documented in this file.
 - Continuous 2 second interval thumbnail comparison
 - Brush thumbnails caching
 
+
+
 ## [1.0.7] - 2025-12-24
 
 ### Changed
 - Reworked `thumbnail_manager.py` to use less CPU resources, previous version caused stuttering during painting
 - 10 brushes are detected for thumbnail changes every 2 seconds instead of 50 brushes
 - 16 x 16 sample points to detect thumbnail changes instead of 32 x 32
+
+
 
 ## [1.0.6] - 2025-12-24
 
@@ -38,6 +57,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - Option to change font size for displayed brush names
+
+
 
 ## [1.0.5] - 2025-12-23
 
@@ -50,10 +71,14 @@ All notable changes to this project will be documented in this file.
 - MD5-based thumbnail hash detection
 - Recognizes separate brush presets being used for eraser and brush
 
+
+
 ## [1.0.4] - 2025-12-23
 
 ### Changed
 - Lowered `setMinimumWidth` in `preset_groups.py` to let users make the docker even smaller for compact setups
+
+
 
 ## [1.0.3] - 2025-12-22
 
@@ -61,10 +86,14 @@ All notable changes to this project will be documented in this file.
 - "Add Brush" icon uniformity. The icon was previously an outlined square and a + symbol which could be offset on different Krita setups
 - Clarification on what "Add Brush" does in the settings dialog
 
+
+
 ## [1.0.2] - 2025-12-22
 
 ### Added
 - Option for exclusive uncollapse mode: only one group can be uncollapsed, uncollapsing a group selects it
+
+
 
 ## [1.0.1] - 2025-12-22